+++ lib/pragma/decorator/association/adapter/active_record.rb
@@ -9,6 +9,8 @@ module Pragma
         #
         # @api private
         class ActiveRecord < Base
+          include Adaptor
+
           class << self
             # Returns whether the adapter supports the given model.
             #