+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Use [Adaptor](https://github.com/aldesantis/adaptor.rb) for association and pagination adpators
+
 ## [2.2.0]
 
 ### Added